requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
assemblyai==0.48.4
azure-core==1.36.0
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
bcrypt==4.1.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
isodate==0.7.2
jiter==0.12.0
mediapipe==0.10.32
moviepy==2.2.1
numpy==2.4.0
opencv-python==4.11.0.86
openai==2.14.0
pillow==11.0.0
proglog==0.1.12
psycopg==3.2.3
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.8.0
python-dotenv==1.2.1
python-multipart==0.0.6
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
websockets==15.0.1
yt-dlp==2025.11.12
pytest==8.3.2
easyocr
